fix: rewrite constraint.txt generation again
Simplify and fix the logic for trying to determine a single version of
every package that meets all of the requirements specifiers for that
package. Add comments explaining the reasoning for each part of the
algorithm. Improve logging for that section of the code to make
debugging easier.

The fix ensures that we examine all packages that want any version of a
dependency and that we track the versions of the dependency that might
be candidates to be used more correctly so that we do not mistakenly
choose a version when it is not actually compatible.